requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
pydantic>=2.6
email-validator>=2.1
python-dotenv>=1.0
boto3>=1.34
python-multipart>=0.0.9
Pillow>=10.2
httpx>=0.27
